requirements.txt
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-ldap==3.4.2
